Update Theoretical_slope.m
Fixed bugs including making sure the index of refraction and wavelength are in the medium around the beads. By computing the nominal values we can test the outputs of computation at the limit where the acceptance angle of the transmissometer is zero, the beads are monospecific, bb wavelength and angular reponse are delta functions.